Update to 2.1.0

+++ src/utils/settings.rb
@@ -37,6 +37,8 @@ require_relative 'compiler'
 DEBUG = false
 INFO = false
 
+SETTINGS_WORDS_BITS_PER_CHAR = 5
+
 def dputs(s)
     puts s if DEBUG
 end
@@ -77,6 +79,7 @@ end
 
 class NameEncoder
     attr_reader :max_length
+    attr_reader :max_word_length
 
     def initialize(names, max_length)
         @names = names
@@ -90,6 +93,8 @@ class NameEncoder
         @non_split = Set.new
         # Key is the name, value is its encoding
         @encoded = Hash.new
+
+        @max_word_length = 0;
 
         update_words
         encode_names
@@ -105,10 +110,14 @@ class NameEncoder
 
     def estimated_size(settings_count)
         size = 0
+        @max_word_length = 0
         @words.each do |word, count|
-            size += word.length + 1
+            size += (word.length + 1) * (5/8.0)
+            if word.length > @max_word_length
+                @max_word_length = word.length
+            end
         end
-        return size + @max_length * settings_count
+        return size.to_i + @max_length * settings_count
     end
 
     def format_encoded_name(name)
@@ -240,8 +249,7 @@ class ValueEncoder
         return buf.to_carr
     end
 
-    private
-    def encode_value(buf, val)
+    def resolve_value(val)
         v = val || 0
         if !v.is_number_kind?
             v = @constants[val]
@@ -249,6 +257,12 @@ class ValueEncoder
                 raise "Could not resolve constant #{val}"
             end
         end
+        return v
+    end
+
+    private
+    def encode_value(buf, val)
+        v = resolve_value(val)
         pos = @values.find_index(v)
         if pos < 0
             raise "Could not encode value not in array #{v}"
@@ -260,10 +274,10 @@ end
 OFF_ON_TABLE = Hash["name" => "off_on", "values" => ["OFF", "ON"]]
 
 class Generator
-    def initialize(src_root, settings_file)
+    def initialize(src_root, settings_file, output_dir)
         @src_root = src_root
         @settings_file = settings_file
-        @output_dir = File.dirname(settings_file)
+        @output_dir = output_dir || File.dirname(settings_file)
 
         @compiler = Compiler.new
 
@@ -324,11 +338,11 @@ class Generator
         puts "each setting name uses #{@name_encoder.max_length} bytes"
         puts "#{@name_encoder.estimated_size(@count)} bytes estimated for setting name storage"
         values_size = @value_encoder.values.length * 4
-        puts "value storage uses #{values_size} bytes"
+        puts "min/max value storage uses #{values_size} bytes"
         value_idx_size = @value_encoder.index_bytes * 2
         value_idx_total = value_idx_size * @count
         puts "value indexing uses #{value_idx_size} per setting, #{value_idx_total} bytes total"
-        puts "#{value_idx_size+value_idx_total} bytes estimated for value storage"
+        puts "#{value_idx_size+value_idx_total} bytes estimated for value+indexes storage"
 
         buf = StringIO.new
         buf << "#include \"fc/settings.h\"\n"
@@ -365,10 +379,12 @@ class Generator
         buf << "#pragma once\n"
         # Write setting_t size constants
         buf << "#define SETTING_MAX_NAME_LENGTH #{@max_name_length+1}\n" # +1 for the terminating '\0'
+        buf << "#define SETTING_MAX_WORD_LENGTH #{@name_encoder.max_word_length+1}\n" # +1 for the terminating '\0'
         buf << "#define SETTING_ENCODED_NAME_MAX_BYTES #{@name_encoder.max_length}\n"
         if @name_encoder.uses_byte_indexing
             buf << "#define SETTING_ENCODED_NAME_USES_BYTE_INDEXING\n"
         end
+        buf << "#define SETTINGS_WORDS_BITS_PER_CHAR #{SETTINGS_WORDS_BITS_PER_CHAR}\n"
         buf << "#define SETTINGS_TABLE_COUNT #{@count}\n"
         offset_type = "uint16_t"
         if can_use_byte_offsetof
@@ -401,7 +417,11 @@ class Generator
         ii = 0
         foreach_enabled_member do |group, member|
             name = member["name"]
-            buf << "#define SETTING_#{name.upcase} #{ii}\n"
+            min, max = resolve_range(member)
+            setting_name = "SETTING_#{name.upcase}"
+            buf << "#define #{setting_name} #{ii}\n"
+            buf << "#define #{setting_name}_MIN #{min}\n"
+            buf << "#define #{setting_name}_MAX #{max}\n"
             ii += 1
         end
 
@@ -416,7 +436,7 @@ class Generator
         }
         add_header.call("platform.h")
         add_header.call("config/parameter_group_ids.h")
-        add_header.call("settings.h")
+        add_header.call("fc/settings.h")
 
         foreach_enabled_group do |group|
             (group["headers"] || []).each do |h|
@@ -450,13 +470,63 @@ class Generator
         buf << "};\n"
 
         # Write word list
-        buf << "static const char *settingNamesWords[] = {\n"
-        buf << "\tNULL,\n"
+        buf << "static const uint8_t settingNamesWords[] = {\n"
+        word_bits = SETTINGS_WORDS_BITS_PER_CHAR
+        # We need 27 symbols for a-z + null
+        rem_symbols = 2 ** word_bits - 27
+        symbols = Array.new
+        acc = 0
+        acc_bits = 0
+        encode_byte = lambda do |c|
+            if c == 0
+                chr = 0 # XXX: Remove this if we go for explicit lengths
+            elsif c >= 'a'.ord && c <= 'z'.ord
+                chr = 1 + (c - 'a'.ord)
+            elsif c >= 'A'.ord && c <= 'Z'.ord
+                raise "Cannot encode uppercase character #{c.ord} (#{c})"
+            else
+                idx = symbols.index(c)
+                if idx.nil?
+                    if rem_symbols == 0
+                        raise "Cannot encode character #{c.ord} (#{c}), no symbols remaining"
+                    end
+                    rem_symbols -= 1
+                    idx = symbols.length
+                    symbols.push(c)
+                end
+                chr = 1 + ('z'.ord - 'a'.ord + 1) + idx
+            end
+            if acc_bits >= (8 - word_bits)
+                # Write
+                remaining = 8 - acc_bits
+                acc |= chr << (remaining - word_bits)
+                buf << "0x#{acc.to_s(16)},"
+                acc = (chr << (8 - (word_bits - remaining))) & 0xff
+            else
+                # Accumulate for next byte
+                acc |= chr << (3 - acc_bits)
+            end
+            acc_bits = (acc_bits + word_bits) % 8
+        end
         @name_encoder.words.each do |w|
-            buf << "\t#{w.inspect},\n"
+            buf << "\t"
+            w.each_byte {|c| encode_byte.call(c)}
+            encode_byte.call(0)
+            buf << " /* #{w.inspect} */ \n"
+        end
+        if acc_bits > 0
+            buf << "\t0x#{acc.to_s(16)},"
+            if acc_bits > (8 - word_bits)
+                buf << "0x00"
+            end
+            buf << "\n"
         end
         buf << "};\n"
 
+        # Output symbol array
+        buf << "static const char wordSymbols[] = {"
+        symbols.each { |s| buf << "'#{s.chr}'," }
+        buf << "};\n"
         # Write the tables
         table_names = ordered_table_names()
         table_names.each do |name|
@@ -468,7 +538,7 @@ class Generator
             buf << "};\n"
         end
 
-        buf << "const lookupTableEntry_t settingLookupTables[] = {\n"
+        buf << "static const lookupTableEntry_t settingLookupTables[] = {\n"
         table_names.each do |name|
             vn = table_variable_name(name)
             buf << "\t{ #{vn}, sizeof(#{vn}) / sizeof(char*) },\n"
@@ -476,7 +546,7 @@ class Generator
         buf << "};\n"
 
         # Write min/max values table
-        buf << "const uint32_t settingMinMaxTable[] = {\n"
+        buf << "static const uint32_t settingMinMaxTable[] = {\n"
         @value_encoder.values.each do |v|
             buf <<  "\t#{v},\n"
         end
@@ -492,7 +562,7 @@ class Generator
         end
 
         # Write setting_t values
-        buf << "const setting_t settingsTable[] = {\n"
+        buf << "static const setting_t settingsTable[] = {\n"
 
         last_group = nil
         foreach_enabled_member do |group, member|
@@ -501,14 +571,19 @@ class Generator
                 buf << "\t// #{group["name"]}\n"
             end
 
-            buf << "\t{ #{@name_encoder.format_encoded_name(member["name"])}, "
+            name = member["name"]
+            buf << "\t{ #{@name_encoder.format_encoded_name(name)}, "
             buf << "#{var_type(member["type"])} | #{value_type(group)}"
             tbl = member["table"]
             if tbl
                 buf << " | MODE_LOOKUP"
                 buf << ", .config.lookup = { #{table_constant_name(tbl)} }"
             else
-                enc = @value_encoder.encode_values(member["min"], member["max"])
+                min, max = resolve_range(member)
+                if min > max
+                    raise "Error encoding #{name}: min (#{min}) > max (#{max})"
+                end
+                enc = @value_encoder.encode_values(min, max)
                 buf <<  ", .config.minmax.indexes = #{enc}"
             end
             buf << ", offsetof(#{group["type"]}, #{member["field"]}) },\n"
@@ -532,6 +607,8 @@ class Generator
             return "VAR_UINT32"
         when "float"
             return "VAR_FLOAT"
+        when "string"
+            return "VAR_STRING"
         else
             raise "unknown variable type #{typ.inspect}"
         end
@@ -539,6 +616,12 @@ class Generator
 
     def value_type(group)
         return group["value_type"] || "MASTER_VALUE"
+    end
+
+    def resolve_range(member)
+        min = @value_encoder.resolve_value(member["min"])
+        max = @value_encoder.resolve_value(member["max"])
+        return min, max
     end
 
     def is_condition_enabled(cond)
@@ -623,17 +706,19 @@ class Generator
         # Use a temporary dir reachable by relative path
         # since g++ in cygwin fails to open files
         # with absolute paths
-        tmp = File.join("obj", "tmp")
+        tmp = File.join(@output_dir, "tmp")
         FileUtils.mkdir_p(tmp) unless File.directory?(tmp)
         value = yield(tmp)
-        FileUtils.remove_dir(tmp)
+        if File.directory?(tmp)
+            FileUtils.remove_dir(tmp)
+        end
         value
     end
 
     def compile_test_file(prog)
         buf = StringIO.new
         # cstddef for offsetof()
-        headers = ["target.h", "platform.h", "cstddef"]
+        headers = ["platform.h", "target.h", "cstddef"]
         @data["groups"].each do |group|
             gh = group["headers"]
             if gh
@@ -651,7 +736,7 @@ class Generator
             file = File.join(dir, "test.cpp")
             File.open(file, 'w') {|file| file.write(buf.string)}
             dputs "Compiling #{buf.string}"
-            stdout, stderr = @compiler.run(file, File.join(dir, "test"))
+            stdout, stderr = @compiler.run(file, File.join(dir, "test"), '-c', noerror: true)
             dputs "Output: #{stderr}"
             stderr
         end
@@ -750,18 +835,22 @@ class Generator
         stderr.scan(/var_(\d+).*?', which is of non-class type '(.*)'/).each do |m|
             member = members[m[0].to_i]
             case m[1]
-            when "int8_t {aka signed char}"
+            when /^int8_t/ # {aka signed char}"
                 typ = "int8_t"
-            when "uint8_t {aka unsigned char}"
+            when /^uint8_t/ # {aka unsigned char}"
                 typ = "uint8_t"
-            when "int16_t {aka short int}"
+            when /^int16_t/ # {aka short int}"
                 typ = "int16_t"
-            when "uint16_t {aka short unsigned int}"
+            when /^uint16_t/ # {aka short unsigned int}"
                 typ = "uint16_t"
-            when "uint32_t {aka long unsigned int}"
+            when /^uint32_t/ # {aka long unsigned int}"
                 typ = "uint32_t"
             when "float"
                 typ = "float"
+            when /^char \[(\d+)\]/
+                # Substract 1 to show the maximum string size without the null terminator
+                member["max"] = $1.to_i - 1;
+                typ = "string"
             else
                 raise "Unknown type #{m[1]} when resolving type for setting #{member["name"]}"
             end
@@ -882,17 +971,20 @@ if __FILE__ == $0
         exit(1)
     end
 
-    gen = Generator.new(src_root, settings_file)
 
     opts = GetoptLong.new(
+        [ "--output-dir", "-o", GetoptLong::REQUIRED_ARGUMENT ],
         [ "--help", "-h", GetoptLong::NO_ARGUMENT ],
         [ "--json", "-j", GetoptLong::REQUIRED_ARGUMENT ],
     )
 
     jsonFile = nil
+    output_dir = nil
 
     opts.each do |opt, arg|
         case opt
+        when "--output-dir"
+            output_dir = arg
         when "--help"
             usage()
             exit(0)
@@ -900,6 +992,8 @@ if __FILE__ == $0
             jsonFile = arg
         end
     end
+
+    gen = Generator.new(src_root, settings_file, output_dir)
 
     if jsonFile
         gen.write_json(jsonFile)
